Added texture coordinate format selection (note: Is this harming Lost World? Check Raptor shadows in-game.)
Emulated transparency in A1RGB5 textures using an alpha test.

Uncommented some code in the Step 1.0 portion of draw_block() (the code which uses the second pointer in a block) which fixes some geometry problems in VF3.

All textures of the unknown type (1 -- probably 4-bit mono) are drawn as bright green to make them stand out.